requirements.txt
aniso8601==7.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
graphene==2.1.8
graphene-file-upload==1.2.2
graphene-mongo==0.2.12
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
passlib==1.7.2
promise==2.3
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
Werkzeug==1.0.1
